% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{Prediction}
\alias{Prediction}
\title{Make prediction based on saved model}
\usage{
Prediction()
}
\arguments{
\item{model}{Model already loaded in environment}
}
\description{
Loads specified model and makes prediction using given predictors
}
\examples{
coming ...
}
\keyword{machine_learning}
\keyword{prediction}
