% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RserveFunctions-package.R
\docType{package}
\name{RserveFunctions-package}
\alias{RserveFunctions}
\alias{RserveFunctions-package}
\title{RserveFunctions: Provides functions for Rserve}
\description{
Defines functions need to provide Machine Learning predictions through a Rserve instance.
}
\author{
\strong{Maintainer}: Claus Bo Hansen \email{sbo@trolling.dk}

}
\keyword{internal}
